backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
networkx>=3.0,<3.2
python-multipart
openai>=1.0.0
pytest
python-dotenv
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
sentence-transformers>=2.2.0
pymupdf
pytesseract
pdf2image
Pillow
